wishlist_project/urls.py
from django.contrib import admin
from django.contrib.auth import views as auth_views
from django.conf.urls.i18n import i18n_patterns
from django.urls import path, include


urlpatterns = [
    path('admin/', admin.site.urls, name='admin'),
    path('accounts/login/', auth_views.LoginView.as_view(), name='login'),
    path('logout/', auth_views.LogoutView.as_view(), name='logout'),
    path('', include('wishlist.urls')),
]

urlpatterns += i18n_patterns(
    path('i18n/setlang/', include('django.conf.urls.i18n')),
)
