requirements.txt
Pillow
psycopg2-binary
django-environ